Validation des fichiers de lettings
